+++ CHANGELOG.md
@@ -127,6 +127,7 @@
 ### Miscellaneous
 
 - Update isoman submodule
+- Update isoman submodule
 - Remove cached dependencies and move all of the tests to this project
 - Update pcm submodule
 - Consolidate common deps into workspace.dependencies